aider: Added a new parameter to control whether to remove the sharing file access or not.
# Aider chat conversation:

USER: Can we add a parameter to remove the expiration which is removing the sharingfileaccess because this is only applied when sharedlink to anyone

ASSISTANT: Sure, we can add a new parameter to control whether to remove the sharing file access or not. This parameter will be a boolean and will default to `true` (meaning the sharing file access will be removed). If the user provides this parameter as `false`, the sharing file access will not be removed.

Here are the changes we need to make:

1. Add a new parameter `$RemoveSharingFileAccess` to the parameters list.
2. Add a condition to check the value of `$RemoveSharingFileAccess` before removing the sharing file access.

Here are the proposed changes:

```powershell
get-sp-rp.ps1
<<<<<<< HEAD
#Parameters
param(
    [Parameter(Mandatory=$true)]
    [string]$SiteUrl,
    [Parameter(Mandatory=$true)]
    [string]$ReportOutput,
    [Parameter(Mandatory=$true)]
    [string]$ListName,
    [Parameter(Mandatory=$false)]
    [string]$LogFile = "log.txt"
)
=======
#Parameters
param(
    [Parameter(Mandatory=$true)]
    [string]$SiteUrl,
    [Parameter(Mandatory=$true)]
    [string]$ReportOutput,
    [Parameter(Mandatory=$true)]
    [string]$ListName,
    [Parameter(Mandatory=$false)]
    [string]$LogFile = "log.txt",
    [Parameter(Mandatory=$false)]
    [bool]$RemoveSharingFileAccess = $true
)
>>>>>>> updated
```

```powershell
get-sp-rp.ps1
<<<<<<< HEAD
#Check if the item is a file or a folder and clear the sharing link accordingly
if ($Item.FileSystemObjectType -eq "File") {
    Remove-PnPFileSharingLink -Url $Item.FieldValues["FileRef"]
} elseif ($Item.FileSystemObjectType -eq "Folder") {
    Remove-PnPFolderSharingLink -Url $Item.FieldValues["FileRef"]
}
=======
#Check if the item is a file or a folder and clear the sharing link accordingly
if ($RemoveSharingFileAccess) {
    if ($Item.FileSystemObjectType -eq "File") {
        Remove-PnPFileSharingLink -Url $Item.FieldValues["FileRef"]
    } elseif ($Item.FileSystemObjectType -eq "Folder") {
        Remove-PnPFolderSharingLink -Url $Item.FieldValues["FileRef"]
    }
}
>>>>>>> updated
```

These changes will add a new parameter `$RemoveSharingFileAccess` and will only remove the sharing file access if this parameter is `true`.

+++ get-sp-rp.ps1
@@ -7,7 +7,9 @@ param(
     [Parameter(Mandatory=$true)]
     [string]$ListName,
     [Parameter(Mandatory=$false)]
-    [string]$LogFile = "log.txt"
+    [string]$LogFile = "log.txt",
+    [Parameter(Mandatory=$false)]
+    [bool]$RemoveSharingFileAccess = $true
 )
     
 try {
@@ -81,10 +83,12 @@ try {
                     }
                      
                     #Check if the item is a file or a folder and clear the sharing link accordingly
-                    if ($Item.FileSystemObjectType -eq "File") {
-                        Remove-PnPFileSharingLink -Url $Item.FieldValues["FileRef"]
-                    } elseif ($Item.FileSystemObjectType -eq "Folder") {
-                        Remove-PnPFolderSharingLink -Url $Item.FieldValues["FileRef"]
+                    if ($RemoveSharingFileAccess) {
+                        if ($Item.FileSystemObjectType -eq "File") {
+                            Remove-PnPFileSharingLink -Url $Item.FieldValues["FileRef"]
+                        } elseif ($Item.FileSystemObjectType -eq "Folder") {
+                            Remove-PnPFolderSharingLink -Url $Item.FieldValues["FileRef"]
+                        }
                     }
 
                     #Collect the data
